Move queries made in Pages/Form/Queries.tsx to Operations.tsx
